+++ src/main/java/com/project/taskmanagement/repository1/PriorityRepository.java
@@ -1,4 +1,4 @@
-package com.project.taskmanagement.repository;
+package com.project.taskmanagement.repository1;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
