+++ src/main/resources/messages.properties
@@ -25,6 +25,8 @@ user.condition.deactivated.code = "It is a deactivated account"
 user.condition.deactivated.message = "The account is in deactivated state"
 user.account.exists.code = "USER_ACCOUNT_EXISTS"
 user.account.exists.message = "User Account Exists"
+user.not.active.code = USER_ID {0} NOT_ACTIVE
+user.not.active.message = User with ID: {0} is not active
 
 priority.exists.code = "PRIORITY_STATUS_ALREADY_EXIST"
 priority.exists.message = "Priority Status already exists"
